Step 10 — account/page.tsx becomes a 110-line router; tabs extracted.
page.tsx was 854 lines with 6 inline tab components, an inline
PasswordStrength widget, a SessionList helper, two redeclared formatDate
forms, a local SocialLinkInput interface that duplicated SocialLink from
shared, and inline-style props strewn through ProfileTab.

New layout:
  components/account/_shared/formatDate.ts    — formatAccountDate
  components/account/PasswordStrength.tsx     — uses shared PASSWORD_RULES
  components/account/SessionList.tsx          — owns its /auth/sessions
                                                 fetch + revoke
  components/account/ProfileTab.tsx           — uses shared User +
                                                 UpdateProfileInput +
                                                 SocialLink; local
                                                 SocialLinkInput deleted;
                                                 SOCIAL_LINK_LIMIT named
                                                 instead of magic 8.
  components/account/SecurityTab.tsx          — composes PasswordStrength
                                                 + SessionList; pulls
                                                 isPasswordValid +
                                                 PASSWORD_RULES from shared
                                                 (placeholder text now
                                                 follows the policy).
  components/account/CommentsTab.tsx          — uses shared UserComment +
                                                 PaginatedResponse.
  components/account/HistoryTab.tsx
  components/account/BookmarksTab.tsx
  components/account/FollowingFeedTab.tsx

page.tsx is a 110-line router: TABS tuple drives both the TabBar and the
conditional render; adding a tab is two lines.

The previous local `const isPasswordValid = sharedIsPasswordValid` alias
goes away — SecurityTab imports it from @derecksnotes/shared directly.

Tested: bunx tsc --noEmit clean on server + client; bun test 45 pass.
