+++ client/src/app/account/page.tsx
@@ -1,30 +1,9 @@
 'use client';
-import React, { useState, useEffect, useCallback } from 'react';
+import React, { useState } from 'react';
 import { useAuth } from '@/context/AuthContext';
-import { api } from '@/utils/api';
-import type {
-  ReadHistoryEntry,
-  SessionInfo,
-  UserComment,
-  PaginatedResponse
-} from '@derecksnotes/shared';
-import {
-  PASSWORD_RULES,
-  isPasswordValid as sharedIsPasswordValid
-} from '@derecksnotes/shared';
-import { toast } from 'sonner';
 import {
   PageContainer,
   Card,
-  CardTitle,
-  Label,
-  Input,
-  TextArea,
-  Button,
-  ButtonRow,
-  InfoRow,
-  InfoLabel,
-  InfoValue,
   Avatar,
   ProfileHeader,
   ProfileInfo,
@@ -32,19 +11,16 @@ import {
   DisplayName,
   JoinDate,
   EmptyState,
-  SuccessMessage,
-  ErrorMessage,
   TabBar,
   Tab
 } from '@/components/ui/PageStyles';
-
-function formatDate(iso: string): string {
-  return new Date(iso).toLocaleDateString('en-US', {
-    year: 'numeric',
-    month: 'long',
-    day: 'numeric'
-  });
-}
+import { ProfileTab } from '@/components/account/ProfileTab';
+import { SecurityTab } from '@/components/account/SecurityTab';
+import { CommentsTab } from '@/components/account/CommentsTab';
+import { HistoryTab } from '@/components/account/HistoryTab';
+import { BookmarksTab } from '@/components/account/BookmarksTab';
+import { FollowingFeedTab } from '@/components/account/FollowingFeedTab';
+import { formatAccountDate } from '@/components/account/_shared/formatDate';
 
 type ActiveTab =
   | 'profile'
@@ -53,6 +29,15 @@ type ActiveTab =
   | 'history'
   | 'bookmarks'
   | 'feed';
+
+const TABS: { key: ActiveTab; label: string }[] = [
+  { key: 'profile', label: 'Profile' },
+  { key: 'security', label: 'Security' },
+  { key: 'comments', label: 'My Comments' },
+  { key: 'history', label: 'Read History' },
+  { key: 'bookmarks', label: 'Bookmarks' },
+  { key: 'feed', label: 'Following' }
+];
 
 export default function AccountPage() {
   const {
@@ -95,30 +80,21 @@ export default function AccountPage() {
           <ProfileInfo>
             <Username>{user.username}</Username>
             {user.displayName && <DisplayName>{user.displayName}</DisplayName>}
-            <JoinDate>Joined {formatDate(user.createdAt)}</JoinDate>
+            <JoinDate>Joined {formatAccountDate(user.createdAt)}</JoinDate>
           </ProfileInfo>
         </ProfileHeader>
       </Card>
 
       <TabBar>
-        <Tab $active={tab === 'profile'} onClick={() => setTab('profile')}>
-          Profile
-        </Tab>
-        <Tab $active={tab === 'security'} onClick={() => setTab('security')}>
-          Security
-        </Tab>
-        <Tab $active={tab === 'comments'} onClick={() => setTab('comments')}>
-          My Comments
-        </Tab>
-        <Tab $active={tab === 'history'} onClick={() => setTab('history')}>
-          Read History
-        </Tab>
-        <Tab $active={tab === 'bookmarks'} onClick={() => setTab('bookmarks')}>
-          Bookmarks
-        </Tab>
-        <Tab $active={tab === 'feed'} onClick={() => setTab('feed')}>
-          Following
-        </Tab>
+        {TABS.map((t) => (
+          <Tab
+            key={t.key}
+            $active={tab === t.key}
+            onClick={() => setTab(t.key)}
+          >
+            {t.label}
+          </Tab>
+        ))}
       </TabBar>
 
       {tab === 'profile' && (
@@ -140,715 +116,5 @@ export default function AccountPage() {
       {tab === 'bookmarks' && <BookmarksTab />}
       {tab === 'feed' && <FollowingFeedTab />}
     </PageContainer>
-  );
-}
-
-// ============================================================================
-// Following Feed tab
-// ============================================================================
-
-interface FeedItem {
-  id: string;
-  content: string;
-  createdAt: string;
-  slug: string;
-  postTitle: string;
-  user: {
-    id: string;
-    username: string;
-    displayName: string | null;
-    avatarUrl: string | null;
-  } | null;
-}
-
-function FollowingFeedTab() {
-  const [items, setItems] = useState<FeedItem[]>([]);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    api
-      .get<{ data: FeedItem[] }>('/users/me/following-feed?limit=30')
-      .then((d) => setItems(d.data))
-      .catch(() => {})
-      .finally(() => setLoading(false));
-  }, []);
-
-  if (loading) {
-    return (
-      <Card>
-        <CardTitle>Following feed</CardTitle>
-        <EmptyState>Loading…</EmptyState>
-      </Card>
-    );
-  }
-  if (items.length === 0) {
-    return (
-      <Card>
-        <CardTitle>Following feed</CardTitle>
-        <EmptyState>
-          Nothing yet — find users to follow and their recent comments will show
-          up here.
-        </EmptyState>
-      </Card>
-    );
-  }
-  return (
-    <Card>
-      <CardTitle>Following feed</CardTitle>
-      {items.map((it) => (
-        <InfoRow key={it.id}>
-          <div>
-            <InfoValue>
-              <a href={`/profile/${it.user?.username}`}>
-                @{it.user?.username || 'unknown'}
-              </a>{' '}
-              commented on <a href={`/${it.slug}`}>{it.postTitle || it.slug}</a>
-            </InfoValue>
-            <br />
-            <InfoLabel>
-              {it.content.length > 200
-                ? it.content.substring(0, 200) + '…'
-                : it.content}
-            </InfoLabel>
-            <br />
-            <InfoLabel>{formatDate(it.createdAt)}</InfoLabel>
-          </div>
-        </InfoRow>
-      ))}
-    </Card>
-  );
-}
-
-// ============================================================================
-// Bookmarks tab
-// ============================================================================
-
-interface BookmarkEntry {
-  id: string;
-  slug: string;
-  title: string;
-  createdAt: string;
-}
-
-function BookmarksTab() {
-  const [items, setItems] = useState<BookmarkEntry[]>([]);
-  const [page, setPage] = useState(1);
-  const [hasMore, setHasMore] = useState(false);
-  const [loading, setLoading] = useState(true);
-
-  const load = useCallback(async (p: number) => {
-    setLoading(true);
-    try {
-      const data = await api.get<{
-        data: BookmarkEntry[];
-        hasMore: boolean;
-      }>(`/users/me/bookmarks?page=${p}&limit=20`);
-      setItems((prev) => (p === 1 ? data.data : [...prev, ...data.data]));
-      setHasMore(data.hasMore);
-      setPage(p);
-    } finally {
-      setLoading(false);
-    }
-  }, []);
-
-  useEffect(() => {
-    load(1);
-  }, [load]);
-
-  if (loading && items.length === 0) {
-    return (
-      <Card>
-        <CardTitle>Saved posts</CardTitle>
-        <EmptyState>Loading…</EmptyState>
-      </Card>
-    );
-  }
-  if (items.length === 0) {
-    return (
-      <Card>
-        <CardTitle>Saved posts</CardTitle>
-        <EmptyState>You haven&apos;t bookmarked anything yet.</EmptyState>
-      </Card>
-    );
-  }
-  return (
-    <Card>
-      <CardTitle>Saved posts</CardTitle>
-      {items.map((b) => (
-        <InfoRow key={b.id}>
-          <div>
-            <InfoValue>
-              <a href={`/${b.slug}`}>{b.title || b.slug}</a>
-            </InfoValue>
-            <br />
-            <InfoLabel>Saved {formatDate(b.createdAt)}</InfoLabel>
-          </div>
-        </InfoRow>
-      ))}
-      {hasMore && (
-        <ButtonRow>
-          <Button $variant="secondary" onClick={() => load(page + 1)}>
-            Load more
-          </Button>
-        </ButtonRow>
-      )}
-    </Card>
-  );
-}
-
-// ============================================================================
-// Profile Tab
-// ============================================================================
-
-interface SocialLinkInput {
-  label: string;
-  url: string;
-}
-
-function ProfileTab({
-  user,
-  updateProfile,
-  changeUsername
-}: {
-  user: {
-    username: string;
-    displayName: string | null;
-    bio: string | null;
-    location?: string | null;
-    socialLinks?: SocialLinkInput[];
-  };
-  updateProfile: (data: {
-    displayName?: string;
-    bio?: string;
-    location?: string | null;
-    socialLinks?: SocialLinkInput[] | null;
-  }) => Promise<void>;
-  changeUsername: (username: string) => Promise<void>;
-}) {
-  const [displayName, setDisplayName] = useState(user.displayName || '');
-  const [bio, setBio] = useState(user.bio || '');
-  const [location, setLocation] = useState(user.location || '');
-  const [socialLinks, setSocialLinks] = useState<SocialLinkInput[]>(
-    user.socialLinks && user.socialLinks.length > 0
-      ? user.socialLinks
-      : [{ label: '', url: '' }]
-  );
-  const [newUsername, setNewUsername] = useState('');
-  const [saving, setSaving] = useState(false);
-  const [success, setSuccess] = useState<string | null>(null);
-  const [error, setError] = useState<string | null>(null);
-
-  const updateLink = (idx: number, patch: Partial<SocialLinkInput>) =>
-    setSocialLinks((prev) =>
-      prev.map((l, i) => (i === idx ? { ...l, ...patch } : l))
-    );
-
-  const addLink = () =>
-    setSocialLinks((prev) =>
-      prev.length < 8 ? [...prev, { label: '', url: '' }] : prev
-    );
-
-  const removeLink = (idx: number) =>
-    setSocialLinks((prev) =>
-      prev.length === 1
-        ? [{ label: '', url: '' }]
-        : prev.filter((_, i) => i !== idx)
-    );
-
-  const handleSaveProfile = async () => {
-    setSaving(true);
-    setSuccess(null);
-    setError(null);
-    try {
-      const cleaned = socialLinks
-        .map((l) => ({ label: l.label.trim(), url: l.url.trim() }))
-        .filter((l) => l.label && l.url);
-      await updateProfile({
-        displayName: displayName || undefined,
-        bio: bio || undefined,
-        location: location || null,
-        socialLinks: cleaned.length > 0 ? cleaned : null
-      });
-      setSuccess('Profile updated');
-    } catch {
-      setError('Failed to update profile (HTTPS URLs only)');
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  const handleChangeUsername = async () => {
-    if (!newUsername.trim()) return;
-    setSaving(true);
-    setSuccess(null);
-    setError(null);
-    try {
-      await changeUsername(newUsername.trim());
-      setNewUsername('');
-      setSuccess('Username changed');
-    } catch (err: any) {
-      setError(err.data?.error || 'Failed to change username');
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  return (
-    <>
-      <Card>
-        <CardTitle>Edit Profile</CardTitle>
-        <Label>Display Name</Label>
-        <Input
-          value={displayName}
-          onChange={(e) => setDisplayName(e.target.value)}
-          placeholder="Your display name"
-          maxLength={50}
-        />
-        <Label>Bio</Label>
-        <TextArea
-          value={bio}
-          onChange={(e) => setBio(e.target.value)}
-          placeholder="Tell us about yourself"
-          maxLength={500}
-        />
-        <Label>Location</Label>
-        <Input
-          value={location}
-          onChange={(e) => setLocation(e.target.value)}
-          placeholder="City, country (optional)"
-          maxLength={100}
-        />
-        <Label>Social links (HTTPS only, up to 8)</Label>
-        {socialLinks.map((l, i) => (
-          <div
-            key={i}
-            style={{ display: 'flex', gap: '0.5rem', marginBottom: '0.25rem' }}
-          >
-            <Input
-              value={l.label}
-              onChange={(e) => updateLink(i, { label: e.target.value })}
-              placeholder="Label (e.g. GitHub)"
-              maxLength={30}
-              style={{ width: '140px' }}
-            />
-            <Input
-              value={l.url}
-              onChange={(e) => updateLink(i, { url: e.target.value })}
-              placeholder="https://…"
-              maxLength={500}
-              style={{ flex: 1 }}
-            />
-            <Button
-              type="button"
-              $variant="secondary"
-              onClick={() => removeLink(i)}
-              style={{ padding: '4px 10px' }}
-            >
-              −
-            </Button>
-          </div>
-        ))}
-        {socialLinks.length < 8 && (
-          <ButtonRow>
-            <Button
-              type="button"
-              $variant="secondary"
-              onClick={addLink}
-              style={{ padding: '4px 10px' }}
-            >
-              + Add link
-            </Button>
-          </ButtonRow>
-        )}
-        {success && <SuccessMessage>{success}</SuccessMessage>}
-        {error && <ErrorMessage>{error}</ErrorMessage>}
-        <ButtonRow>
-          <Button onClick={handleSaveProfile} disabled={saving}>
-            {saving ? 'Saving...' : 'Save'}
-          </Button>
-        </ButtonRow>
-      </Card>
-
-      <Card>
-        <CardTitle>Change Username</CardTitle>
-        <Label>New Username</Label>
-        <Input
-          value={newUsername}
-          onChange={(e) => setNewUsername(e.target.value)}
-          placeholder="New username (3-30 chars)"
-          maxLength={30}
-        />
-        <ButtonRow>
-          <Button
-            onClick={handleChangeUsername}
-            disabled={saving || !newUsername.trim()}
-          >
-            Change Username
-          </Button>
-        </ButtonRow>
-      </Card>
-    </>
-  );
-}
-
-// ============================================================================
-// Password Helpers
-// ============================================================================
-
-const isPasswordValid = sharedIsPasswordValid;
-
-function PasswordStrength({ password }: { password: string }) {
-  const checks = [
-    {
-      label: `${PASSWORD_RULES.minLength}+ characters`,
-      pass: password.length >= PASSWORD_RULES.minLength
-    },
-    { label: 'Uppercase letter', pass: /[A-Z]/.test(password) },
-    { label: 'Lowercase letter', pass: /[a-z]/.test(password) },
-    { label: 'Number', pass: /[0-9]/.test(password) }
-  ];
-
-  return (
-    <div
-      style={{
-        fontSize: '0.75rem',
-        margin: '0 0 0.5rem',
-        display: 'flex',
-        gap: '0.75rem',
-        flexWrap: 'wrap'
-      }}
-    >
-      {checks.map((c) => (
-        <span key={c.label} style={{ color: c.pass ? '#0F9960' : '#999' }}>
-          {c.pass ? '\u2713' : '\u2717'} {c.label}
-        </span>
-      ))}
-    </div>
-  );
-}
-
-// ============================================================================
-// Security Tab
-// ============================================================================
-
-function SecurityTab({
-  changePassword,
-  deleteAccount,
-  logout
-}: {
-  changePassword: (current: string, newPass: string) => Promise<void>;
-  deleteAccount: () => Promise<void>;
-  logout: () => Promise<void>;
-}) {
-  const [currentPassword, setCurrentPassword] = useState('');
-  const [newPassword, setNewPassword] = useState('');
-  const [saving, setSaving] = useState(false);
-  const [success, setSuccess] = useState<string | null>(null);
-  const [error, setError] = useState<string | null>(null);
-
-  const handleChangePassword = async () => {
-    setSaving(true);
-    setSuccess(null);
-    setError(null);
-    try {
-      await changePassword(currentPassword, newPassword);
-      setCurrentPassword('');
-      setNewPassword('');
-      setSuccess('Password changed. All other sessions revoked.');
-    } catch (err: any) {
-      setError(err.data?.error || 'Failed to change password');
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  const handleDeleteAccount = async () => {
-    if (
-      !confirm(
-        'Are you sure you want to delete your account? This cannot be undone.'
-      )
-    )
-      return;
-    if (!confirm('Really delete your account?')) return;
-    try {
-      await deleteAccount();
-    } catch {
-      setError('Failed to delete account');
-    }
-  };
-
-  return (
-    <>
-      <Card>
-        <CardTitle>Change Password</CardTitle>
-        <Label>Current Password</Label>
-        <Input
-          type="password"
-          value={currentPassword}
-          onChange={(e) => setCurrentPassword(e.target.value)}
-        />
-        <Label>New Password</Label>
-        <Input
-          type="password"
-          value={newPassword}
-          onChange={(e) => setNewPassword(e.target.value)}
-          placeholder="Min 8 chars, 1 upper, 1 lower, 1 number"
-        />
-        {newPassword && <PasswordStrength password={newPassword} />}
-        {success && <SuccessMessage>{success}</SuccessMessage>}
-        {error && <ErrorMessage>{error}</ErrorMessage>}
-        <ButtonRow>
-          <Button
-            onClick={handleChangePassword}
-            disabled={
-              saving ||
-              !currentPassword ||
-              !newPassword ||
-              !isPasswordValid(newPassword)
-            }
-          >
-            {saving ? 'Changing...' : 'Change Password'}
-          </Button>
-        </ButtonRow>
-      </Card>
-
-      <Card>
-        <CardTitle>Sessions</CardTitle>
-        <SessionList />
-      </Card>
-
-      <Card>
-        <CardTitle>Danger Zone</CardTitle>
-        <ButtonRow>
-          <Button $variant="secondary" onClick={logout}>
-            Log Out Everywhere
-          </Button>
-          <Button $variant="danger" onClick={handleDeleteAccount}>
-            Delete Account
-          </Button>
-        </ButtonRow>
-      </Card>
-    </>
-  );
-}
-
-function SessionList() {
-  const [sessions, setSessions] = useState<SessionInfo[]>([]);
-
-  useEffect(() => {
-    api
-      .get<SessionInfo[]>('/auth/sessions')
-      .then(setSessions)
-      .catch(() => {
-        toast.error('Failed to load sessions');
-      });
-  }, []);
-
-  const revoke = async (id: string) => {
-    try {
-      await api.delete(`/auth/sessions/${id}`);
-      setSessions((prev) => prev.filter((s) => s.id !== id));
-      toast.success('Session revoked');
-    } catch {
-      toast.error('Failed to revoke session');
-    }
-  };
-
-  return (
-    <>
-      {sessions.map((s) => (
-        <InfoRow key={s.id}>
-          <div>
-            <InfoValue>
-              {s.isCurrent
-                ? 'Current session'
-                : s.userAgent?.substring(0, 40) || 'Unknown device'}
-            </InfoValue>
-            <br />
-            <InfoLabel>
-              {s.ipAddress} — expires {formatDate(s.expiresAt)}
-            </InfoLabel>
-          </div>
-          {!s.isCurrent && (
-            <Button $variant="secondary" onClick={() => revoke(s.id)}>
-              Revoke
-            </Button>
-          )}
-        </InfoRow>
-      ))}
-      {sessions.length === 0 && <EmptyState>No active sessions.</EmptyState>}
-    </>
-  );
-}
-
-// ============================================================================
-// Comments Tab
-// ============================================================================
-
-function CommentsTab() {
-  const [comments, setComments] = useState<UserComment[]>([]);
-  const [selected, setSelected] = useState<Set<string>>(new Set());
-  const [page, setPage] = useState(1);
-  const [hasMore, setHasMore] = useState(false);
-
-  useEffect(() => {
-    loadComments(1);
-  }, []);
-
-  const loadComments = async (p: number) => {
-    const data = await api.get<PaginatedResponse<UserComment>>(
-      `/users/me/comments?page=${p}&limit=20`
-    );
-    if (p === 1) setComments(data.data);
-    else setComments((prev) => [...prev, ...data.data]);
-    setHasMore(data.hasMore);
-    setPage(p);
-  };
-
-  const toggle = (id: string) => {
-    setSelected((prev) => {
-      const next = new Set(prev);
-      if (next.has(id)) next.delete(id);
-      else next.add(id);
-      return next;
-    });
-  };
-
-  const bulkDelete = async () => {
-    if (!confirm(`Delete ${selected.size} comment(s)?`)) return;
-    try {
-      await api.post('/users/me/comments/bulk-delete', {
-        commentIds: Array.from(selected)
-      });
-      toast.success(`${selected.size} comment(s) deleted`);
-      setSelected(new Set());
-      loadComments(1);
-    } catch {
-      toast.error('Failed to delete comments');
-    }
-  };
-
-  return (
-    <Card>
-      <CardTitle>My Comments</CardTitle>
-      {selected.size > 0 && (
-        <ButtonRow>
-          <Button $variant="danger" onClick={bulkDelete}>
-            Delete Selected ({selected.size})
-          </Button>
-          <Button $variant="secondary" onClick={() => setSelected(new Set())}>
-            Clear Selection
-          </Button>
-        </ButtonRow>
-      )}
-      {comments.length === 0 ? (
-        <EmptyState>No comments yet.</EmptyState>
-      ) : (
-        comments.map((c) => (
-          <InfoRow key={c.id}>
-            <div
-              style={{
-                display: 'flex',
-                gap: '0.5rem',
-                alignItems: 'flex-start',
-                flex: 1
-              }}
-            >
-              <input
-                type="checkbox"
-                checked={selected.has(c.id)}
-                onChange={() => toggle(c.id)}
-                disabled={c.isDeleted}
-              />
-              <div style={{ flex: 1 }}>
-                <InfoValue
-                  style={{
-                    textDecoration: c.isDeleted ? 'line-through' : 'none'
-                  }}
-                >
-                  {c.content.substring(0, 100)}
-                  {c.content.length > 100 ? '...' : ''}
-                </InfoValue>
-                <br />
-                <InfoLabel>
-                  on <a href={`/${c.slug}`}>{c.postTitle || c.slug}</a> —{' '}
-                  {formatDate(c.createdAt)}
-                  {!c.approved && ' — pending'}
-                  {c.isDeleted && ' — deleted'}
-                </InfoLabel>
-              </div>
-            </div>
-            <div
-              style={{
-                display: 'flex',
-                gap: '0.5rem',
-                fontSize: '0.75rem'
-              }}
-            >
-              <span title="Likes">+{c.likes}</span>
-              <span title="Dislikes">-{c.dislikes}</span>
-            </div>
-          </InfoRow>
-        ))
-      )}
-      {hasMore && (
-        <ButtonRow>
-          <Button $variant="secondary" onClick={() => loadComments(page + 1)}>
-            Load More
-          </Button>
-        </ButtonRow>
-      )}
-    </Card>
-  );
-}
-
-// ============================================================================
-// History Tab
-// ============================================================================
-
-function HistoryTab() {
-  const [entries, setEntries] = useState<ReadHistoryEntry[]>([]);
-  const [page, setPage] = useState(1);
-  const [hasMore, setHasMore] = useState(false);
-
-  useEffect(() => {
-    loadHistory(1);
-  }, []);
-
-  const loadHistory = async (p: number) => {
-    const data = await api.get<PaginatedResponse<ReadHistoryEntry>>(
-      `/users/me/read-history?page=${p}&limit=20`
-    );
-    if (p === 1) setEntries(data.data);
-    else setEntries((prev) => [...prev, ...data.data]);
-    setHasMore(data.hasMore);
-    setPage(p);
-  };
-
-  return (
-    <Card>
-      <CardTitle>Read History</CardTitle>
-      {entries.length === 0 ? (
-        <EmptyState>No reading history yet.</EmptyState>
-      ) : (
-        entries.map((e, i) => (
-          <InfoRow key={i}>
-            <div>
-              <InfoValue>
-                <a href={`/${e.postSlug}`}>{e.postTitle || e.postSlug}</a>
-              </InfoValue>
-            </div>
-            <InfoLabel>{formatDate(e.readAt)}</InfoLabel>
-          </InfoRow>
-        ))
-      )}
-      {hasMore && (
-        <ButtonRow>
-          <Button $variant="secondary" onClick={() => loadHistory(page + 1)}>
-            Load More
-          </Button>
-        </ButtonRow>
-      )}
-    </Card>
   );
 }
