+++ client/src/components/account/_shared/formatDate.ts
@@ -1,0 +1,7 @@
+export function formatAccountDate(iso: string): string {
+  return new Date(iso).toLocaleDateString('en-US', {
+    year: 'numeric',
+    month: 'long',
+    day: 'numeric'
+  });
+}
